% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\docType{data}
\name{ae}
\alias{ae}
\title{SDTM Adverse Events data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{AESEQ}}{[double] Sequence Number}
\item{\code{AESTDT}}{[double] Start date of the adverse event}
\item{\code{AESTDY}}{[double] The number of days the subject had been enrolled in the study when the adverse event began.}
\item{\code{AEENDT}}{[double] End date of the adverse event}
\item{\code{AEENDY}}{[double] The number of days the subject had been enrolled in the study when the adverse event ended.}
\item{\code{AETERM}}{[character] The reported term for the adverse event, i.e. how the reporter described the adverse event.}
\item{\code{AEDECOD}}{[character] The official term for the adverse event, i.e. the dictionary derived description of the adverse event.}
\item{\code{AEBODSYS}}{[character] Body system or organ class involved in the adverse event.}
\item{\code{AESER}}{[character] Was the adverse event serious? \code{Y} = Yes, \code{N} = No.}
\item{\code{AEONGO}}{[character] Is the adverse event ongoing? \code{Y} = Yes, \code{N} = No.}
\item{\code{AESEV}}{[character] Severity of the adverse event.}
\item{\code{AEREL}}{[character] The investigator's opinion as to whether or not the adverse event was related to the study treatment.}
\item{\code{AEOUT}}{[character] Outcome of the adverse event.}
}
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
ae
}
\description{
\code{ae} describes adverse event data related to a simulated
clinical trial. An adverse event is an undesirable medical occurrence that
happens while a subject is enrolled in a clinical trial. These occurrences
must be reported and analyzed to see if they are caused by the treatment
under study.
}
\concept{Rho SDTM datasets}
\keyword{datasets}
