% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_electricity.R
\docType{data}
\name{aus_electricity}
\alias{aus_electricity}
\alias{aus_electricity_dictionary}
\title{Australian electricity demand data}
\format{
A tibble with 1155264 rows and 3 variables:
\describe{

\item{\code{date}}{[dttm] Starting date-time of demand reading}
\item{\code{state}}{[chr] State abbreviation: Victoria (VIC), New South Wales (NSW), Queensland (QUN), Tasmania (TAS), and South Australia (SA)}
\item{\code{demand}}{[dbl] Half-hourly electricity demand in MW}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4659727}
}
\usage{
aus_electricity

aus_electricity_dictionary
}
\description{
This dataset records half-hourly electricity demand for five states in
Australia: Victoria, New South Wales, Queensland, Tasmania, and South Australia.
}
\section{Functions}{
\itemize{
\item \code{aus_electricity_dictionary}: The \code{aus_electricity} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
aus_electricity \%>\%
  as_tsibble(key = state, index = date)

}
\keyword{datasets}
