% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoin.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\alias{bitcoin_dictionary}
\title{bitcoin}
\format{
A tbl_ts:
\describe{

\item{\code{timestamp}}{[date] Day}
\item{\code{series}}{[fct] Series name}
\item{\code{value}}{[dbl] Value of the series during the specified month}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://forecastingdata.org/}
}
\usage{
bitcoin

bitcoin_dictionary
}
\description{
This dataset contains the potential influencers of the bitcoin price.
There are a total of 18 daily time-series including hash rate, block size, mining difficulty etc.
It also encompasses public opinion in the form of tweets and google searches mentioning the keyword bitcoin.
The data is scraped from the interactive web-graphs available at \url{https://bitinfocharts.com}.

For more details, please refer to
BitInfoCharts, 2021. Cryptocurrency statistics. URL \url{https://bitinfocharts.com}
}
\section{Functions}{
\itemize{
\item \code{bitcoin_dictionary}: The \code{bitcoin} data dictionary
}}

\keyword{datasets}
