% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_parts.R
\docType{data}
\name{car_parts}
\alias{car_parts}
\alias{car_parts_dictionary}
\title{Monthly Car Part Sales}
\format{
A tibble with 136374 rows and 3 variables:
\describe{

\item{\code{part_num}}{[chr] ID of the car part}
\item{\code{date}}{[date] Start date of the month}
\item{\code{qty}}{[dbl] Number of parts sold that month}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656022#.YSzxsNNKj6O}
}
\usage{
car_parts

car_parts_dictionary
}
\description{
This dataset contains 2674 intermittent monthly time series that represent
car parts sales from January 1998 to March 2002.
It was extracted from R expsmooth package.
}
\section{Functions}{
\itemize{
\item \code{car_parts_dictionary}: The \code{car_parts} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
car_parts \%>\%
  as_tsibble(key = part_num, index = date)
}
\keyword{datasets}
