% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\docType{data}
\name{covid}
\alias{covid}
\alias{covid_dictionary}
\title{COVID-19 US Historical Data by State}
\format{
A tibble with 20780 rows and 6 variables:
\describe{

\item{\code{date}}{[double] }
\item{\code{state}}{[integer] Two-letter abbreviation for the state or territory.}
\item{\code{tests}}{[double] Daily increase in totalTestResults, calculated from the previous day’s value.}
\item{\code{cases}}{[double] The daily increase in API field positive, which measures Cases (confirmed plus probable) calculated based on the previous day’s value.}
\item{\code{hospitalizations}}{[double] Daily increase in hospitalizedCumulative, calculated from the previous day’s value.}
\item{\code{deaths}}{[double] Daily increase in death, calculated from the previous day’s value.}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://covidtracking.com/data/download/all-states-history.csv}
}
\usage{
covid

covid_dictionary
}
\description{
From \href{https://covidtracking.com}{The COVID Tracking Project}:

We collect, cross-check, and publish COVID-19
\href{https://covidtracking.com/data}{data} from 56 US states and territories in
three main areas: \strong{testing, hospitalization, and patient outcomes},
\strong{racial and ethnic demographic information} via
\href{https://covidtracking.com/race}{The COVID Racial Data Tracker}, and
\strong{long-term-care facilities} via the
\href{https://covidtracking.com/data/longtermcare}{Long-Term-Care tracker}.
We compile these numbers to provide the most complete picture we can assemble
of the US COVID-19 testing effort and the outbreak’s effects on the people
and communities it strikes.

If you’d like to use the data, whether it’s for a specialized project or just
to better understand COVID-19 in the US, here are a few things you should
know right away.
\itemize{
\item We update \href{https://covidtracking.com/data/}{the full dataset} each day
between about 5:30pm and 7pm Eastern time, with limited additional updates
as new information arrives.
\item All our data comes from state and territory public health authorities or
official statements from state officials. Not all states report all data,
which means we can’t, either. You can
\href{https://covidtracking.com/data-sources}{read more about our data sources here}.
}
}
\section{Functions}{
\itemize{
\item \code{covid_dictionary}: The \code{covid} data dictionary
}}

\seealso{
Other COVID-19 datasets: 
\code{\link{covid_state_pop}}
}
\concept{COVID-19 datasets}
\keyword{datasets}
