% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\docType{data}
\name{covid_state_pop}
\alias{covid_state_pop}
\title{COVID-19 US State Populations}
\format{
A tibble with 50 rows and 2 variables:
\describe{
\item{\code{state}}{[factor] Two-letter abbreviation for the state or territory.}
\item{\code{population}}{[int] The state population, as reported by the COVID-19 Tracking Project.}
}
}
\source{
\url{https://api.covidtracking.com/v2/states.json}
}
\usage{
covid_state_pop
}
\description{
State populations as reported by the
\href{https://covidtracking.com}{The COVID Tracking Project}.
}
\seealso{
Other COVID-19 datasets: 
\code{\link{covid}}
}
\concept{COVID-19 datasets}
\keyword{datasets}
