% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{describe_dataset}
\alias{describe_dataset}
\title{Describe the dataset}
\usage{
describe_dataset(x, ..., .title = NULL)
}
\arguments{
\item{x}{The data set}

\item{...}{Descriptions of column names.}
}
\description{
An internal function to construct a machine-friendly description of the
dataset.
}
\keyword{internal}
