% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\docType{data}
\name{dm}
\alias{dm}
\alias{dm_dictionary}
\title{SDTM Demographics data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{SITE}}{[character] Name of study site}
\item{\code{SITEID}}{[character] Study Site Identifier}
\item{\code{AGE}}{[double] Age of subject}
\item{\code{SEX}}{[character] Sex of subject}
\item{\code{RACE}}{[character] Race of subject}
\item{\code{ARM}}{[character] Treatment arm that subject is assigned to}
\item{\code{ARMCD}}{[character] Code for treatment arm that subject is assigned to}
\item{\code{SBJTSTAT}}{[character] Subject status in study}
\item{\code{RFSTDTC}}{[double] Reference start date. Usually the day the subject begins treatment.}
\item{\code{RFENDTC}}{[double] Reference end date. Usually the day the subject takes their last treatment.}
\item{\code{RFENDY}}{[double] The number of days after the RFSTDTC that the RFENDTC occurred, i.e. the number of days the subject spent in the study/in treatment}
\item{\code{SAFFL}}{[character] Safety population flag. Did the subject actually receive a treatment (including placebo) and should therefore be included in the population of subjects studied to determine the safety of the treatment? \code{Y} = Yes and \code{N} = No}
\item{\code{SAFFN}}{[double] Numeric coding of safety population flag. \code{Y} = Yes and \code{N} = No}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
dm

dm_dictionary
}
\description{
\code{dm} describes demographic data related to subjects of a
simulated clinical trial.
}
\section{Functions}{
\itemize{
\item \code{dm_dictionary}: The \code{dm} data dictionary
}}

\concept{Rho SDTM datasets}
\keyword{datasets}
