% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominick.R
\docType{data}
\name{dominick}
\alias{dominick}
\alias{dominick_dictionary}
\title{Dominick's Finer Foods department sales data}
\format{
A tibble with 279519 rows and 25 variables:
\describe{

\item{\code{store}}{[int] Store number}
\item{\code{date}}{[date] Date}
\item{\code{custcoun}}{[int] Number of customers}
\item{\code{grocery}}{[dbl] Non-specialty grocery sales in dollars}
\item{\code{gm}}{[dbl] General merchandising sales in dollars}
\item{\code{dairy}}{[dbl] Dairy sales in dollars}
\item{\code{frozen}}{[dbl] Frozen food sales in dollars}
\item{\code{meat}}{[dbl] Meat sales in dollars}
\item{\code{fish}}{[dbl] Fish sales in dollars}
\item{\code{produce}}{[dbl] Produce sales in dollars}
\item{\code{saladbar}}{[dbl] Salad bar sales in dollars}
\item{\code{floral}}{[dbl] Floral sales in dollars}
\item{\code{deli}}{[dbl] Deli sales in dollars}
\item{\code{cheese}}{[dbl] Cheese case sales in dollars}
\item{\code{bakery}}{[dbl] Bakery sales in dollars}
\item{\code{pharmacy}}{[dbl] Pharmacy sales in dollars}
\item{\code{jewelry}}{[dbl] Jewelry sales in dollars}
\item{\code{cosmetic}}{[dbl] }
\item{\code{haba}}{[dbl] Health and beauty aids sales in dollars}
\item{\code{camera}}{[dbl] Camera sales in dollars}
\item{\code{photofin}}{[dbl] Photo development sales in dollars}
\item{\code{video}}{[dbl] Video sales in dollars}
\item{\code{beer}}{[dbl] Beer sales in dollars}
\item{\code{wine}}{[dbl] Wine sales in dollars}
\item{\code{spirits}}{[dbl] Alcoholic spirits sales in dollars}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://www.chicagobooth.edu/research/kilts/datasets/dominicks}
}
\usage{
dominick

dominick_dictionary
}
\description{
From the \href{https://www.chicagobooth.edu/research/kilts/datasets/dominicks}{Chicago Booth Kilts Center for Marketing}:

From 1989 to 1994, Chicago Booth and Dominick’s Finer Foods entered into a
partnership for store-level research into shelf management and pricing.
Randomized experiments were conducted in more than 25 different categories
throughout all stores in this 100-store chain. As a by-product of this
research cooperation, approximately nine years of store-level data on the
sales of more than 3,500 UPCs is available in this dataset. This data is
unique for the breadth of its coverage and for the information available on
retail margins.

The customer count file includes information about in-store traffic. The data
is store specific and on a daily basis. The customer count data refers to the
number of customers visiting the store and purchasing something. Also in the
customer count file is a total dollar sales and total coupons redeemed
figure, by DFF defined department. These figures are compiled daily from the
register/scanner receipts.
}
\section{Functions}{
\itemize{
\item \code{dominick_dictionary}: The \code{dominick} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
dominick \%>\%
  as_tsibble(key = c("store"), index = date)
}
\seealso{
Other Dominick's datasets: 
\code{\link{dominick_oatmeal}},
\code{\link{dominick_soap}}
}
\concept{Dominick's datasets}
\keyword{datasets}
