% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominick.R
\docType{data}
\name{dominick_soap}
\alias{dominick_soap}
\alias{dominick_soap_dictionary}
\title{Dominick's Finer Foods bath soap sales data}
\format{
A tibble with 415833 rows and 7 variables:
\describe{

\item{\code{week}}{[date] Start date of the week}
\item{\code{store}}{[int] Store number}
\item{\code{product}}{[chr] Abbreviated product name}
\item{\code{size}}{[chr] Product size}
\item{\code{price}}{[dbl] Price per unit in dollars}
\item{\code{profit}}{[dbl] Profit per unit in dollars}
\item{\code{move}}{[int] Number of units sold during the week}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://www.chicagobooth.edu/research/kilts/datasets/dominicks}
}
\usage{
dominick_soap

dominick_soap_dictionary
}
\description{
From the \href{https://www.chicagobooth.edu/research/kilts/datasets/dominicks}{Chicago Booth Kilts Center for Marketing}:

From 1989 to 1994, Chicago Booth and Dominick’s Finer Foods entered into a
partnership for store-level research into shelf management and pricing.
Randomized experiments were conducted in more than 25 different categories
throughout all stores in this 100-store chain. As a by-product of this
research cooperation, approximately nine years of store-level data on the
sales of more than 3,500 UPCs is available in this dataset. This data is
unique for the breadth of its coverage and for the information available on
retail margins.

The data contain a description of each UPC in a category and sales
information at the store level for each UPC in a category. The information is
stored on a weekly basis.

Note: This is historical data and the products are not for sale.
}
\section{Functions}{
\itemize{
\item \code{dominick_soap_dictionary}: The \code{dominick_soap} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
dominick_soap \%>\%
  as_tsibble(key = c("store", "product", "size", "price"), index = week)
}
\seealso{
Other Dominick's datasets: 
\code{\link{dominick_oatmeal}},
\code{\link{dominick}}
}
\concept{Dominick's datasets}
\keyword{datasets}
