% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec_demand.R
\docType{data}
\name{elec_demand}
\alias{elec_demand}
\alias{elec_demand_dictionary}
\title{Victoria, Australia Electricity Demand Data}
\format{
A tibble with 17520 rows and 2 variables:
\describe{

\item{\code{timestamp}}{[dttm] Datetime of observation}
\item{\code{demand}}{[dbl] Electricity demand (GW)}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656069#.YSe0vdNKiM8}
}
\usage{
elec_demand

elec_demand_dictionary
}
\description{
Single time series representing the half hourly electricity demand (in gigawatts) for
Victoria, Australia in 2014.
}
\section{Functions}{
\itemize{
\item \code{elec_demand_dictionary}: The \code{elec_demand} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
elec_demand \%>\%
  as_tsibble(index = timestamp)
}
\keyword{datasets}
