% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity_weekly.R
\docType{data}
\name{electricity_weekly}
\alias{electricity_weekly}
\alias{electricity_weekly_dictionary}
\title{Weekly Electricity Consumption}
\format{
A tibble with 50076 rows and 3 variables:
\describe{

\item{\code{client}}{[chr] ID of the electric company client}
\item{\code{date}}{[date] Date}
\item{\code{power}}{[dbl] Weekly electricity consumption, in kilowatts (kW)}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656141#.YSkxU9NKiM8}
}
\usage{
electricity_weekly

electricity_weekly_dictionary
}
\description{
This dataset contains weekly aggregated electricity consumption for 321
clients in Portugal from 2012 to 2014.
}
\section{Functions}{
\itemize{
\item \code{electricity_weekly_dictionary}: The \code{electricity_weekly} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
electricity_weekly \%>\%
  as_tsibble(key = client, index = date)
}
\keyword{datasets}
