% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfda.R
\docType{data}
\name{fda_adverse_daily}
\alias{fda_adverse_daily}
\alias{fda_adverse_daily_dictionary}
\title{Daily counts of FDA drug adverse event reports}
\format{
A tibble with 5968 rows and 3 variables:
\describe{

\item{\code{receive_date}}{[date] Date that the report was first received by FDA.}
\item{\code{public}}{[dbl] Number of reports that were submitted directly by a member of the public.}
\item{\code{manufacturer}}{[dbl] Number of reports that were submitted through a drug manufacturer.}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://open.fda.gov/apis/drug/event/}
}
\usage{
fda_adverse_daily

fda_adverse_daily_dictionary
}
\description{
This data set contains daily counts of reports received by the Food and Drug
Administration (FDA) from 2004 to 2020 regarding adverse events associated
with the administration of drugs in medical settings. This data was collected
from the \href{https://open.fda.gov/data/faers/}{FDA Adverse Event Reporting System (FAERS)},
and has been made available through \href{https://open.fda.gov/}{openFDA}.

According to openFDA: "an adverse event is submitted to the FDA to report any
undesirable experience associated with the use of a medical product in a
patient. For drugs, this includes serious drug side effects, product use errors,
product quality problems, and therapeutic failures for prescription or
over-the-counter medicines and medicines administered to hospital patients
or at outpatient infusion centers.

Reporting of adverse events by healthcare professionals and consumers
is voluntary in the United States. FDA receives some adverse event reports
directly from healthcare professionals (such as physicians, pharmacists,
nurses and others) and consumers (such as patients, family members, lawyers
and others). Healthcare professionals and consumers may also report adverse
events to the products’ manufacturers. If a manufacturer receives an adverse
event report, it is normally required to send the report to FDA."
}
\section{Functions}{
\itemize{
\item \code{fda_adverse_daily_dictionary}: The \code{fda_adverse_daily} data dictionary

}}
\seealso{
Other openFDA datasets: 
\code{\link{fda_pt_drugs}}
}
\concept{openFDA datasets}
\keyword{datasets}
