% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfda.R
\docType{data}
\name{fda_pt_drugs}
\alias{fda_pt_drugs}
\alias{fda_pt_drugs_dictionary}
\title{Patient and drug information for FDA drug adverse events}
\format{
A tibble with 5765 rows and 17 variables:
\describe{

\item{\code{report_id}}{[chr] The 8-digit Safety Report ID number, also known as the case report number or case ID. Can be used to identify or find a specific adverse event report.}
\item{\code{receive_date}}{[date] Date that the report was first received by FDA.}
\item{\code{receipt_date}}{[date] Date that the most recent information in the report was received by FDA.}
\item{\code{country}}{[chr] The name of the country where the adverse event occurred.}
\item{\code{reporter}}{[chr] Category of individual who submitted the report: physician, pharmacist, other health professional, laywer or consumer/non-health professional.}
\item{\code{age}}{[dbl] Age of the patient when the adverse event first occured.}
\item{\code{sex}}{[chr] The sex of the patient.}
\item{\code{weight}}{[dbl] The patient weight, in kilograms (kg).}
\item{\code{drug}}{[chr] Drug name. This may be the valid trade name of the product (e.g. "advil" or "aleve") or the generic name (e.g. "ibuprofen").}
\item{\code{dosage}}{[dbl] The number portion of a dosage; when combined with \code{dosage_unit} the complete dosage information is represented.}
\item{\code{dosage_unit}}{[chr] The drug dosasge unit: kilograms (kg), grams (g), milligrams (mg) or micrograms (ug).}
\item{\code{indication}}{[chr] Indication for the drug’s use.}
\item{\code{drug_start_date}}{[date] Date the patient began taking the drug.}
\item{\code{drug_end_date}}{[date] Date the patient stopped taking the drug.}
\item{\code{serious}}{[lgl] A logical value indicating whether or not the adverse event was serious, i.e. resulted in death, a life threatening condition, hospitalization, disability, congenital anomaly, or some other serious condition.}
\item{\code{reaction}}{[chr] Patient reaction, as a term from the Medical Dictionary for Regulatory Activities, encoded in British English.}
\item{\code{outcome}}{[chr] Outcome of the patient reaction at the time of last observation: recovered, recovering, not recovered, recovered with sequelae (consequent health issues), fatal or unknown.}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://open.fda.gov/apis/drug/event/}
}
\usage{
fda_pt_drugs

fda_pt_drugs_dictionary
}
\description{
This data set contains selected variables from a subset of FDA drug adverse
event reports received by the FDA in January 2019.
}
\section{Functions}{
\itemize{
\item \code{fda_pt_drugs_dictionary}: The \code{fda_pt_drugs} data dictionary

}}
\seealso{
Other openFDA datasets: 
\code{\link{fda_adverse_daily}}
}
\concept{openFDA datasets}
\keyword{datasets}
