% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhv.R
\docType{data}
\name{fhv}
\alias{fhv}
\alias{fhv_dictionary}
\title{For-hire Vehicle (FHV) Trips}
\format{
A tibble with 22012 rows and 5 variables:
\describe{

\item{\code{date}}{[date] Date}
\item{\code{base_number}}{[chr] Unique ID of the TLC base company}
\item{\code{base_name}}{[chr] Name of the TLC base company}
\item{\code{trips}}{[dbl] Number of trips}
\item{\code{vehicles}}{[dbl] Number of vehicles available}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/5122535#.YSk1SNNKiDE}
}
\usage{
fhv

fhv_dictionary
}
\description{
This dataset contains number of trips and vehicles belonging to
171 for-hire vehicle (FHV) companies in
New York City from January to August, 2015. Data were originally obtained
from the NYC Taxi & Limousine Commission (TLC).
}
\section{Functions}{
\itemize{
\item \code{fhv_dictionary}: The \code{fhv} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
library(dplyr)
fhv \%>\%
  tsibble::as_tsibble(key = starts_with("base"), index = date)
}
\keyword{datasets}
