% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu.R
\docType{data}
\name{flu}
\alias{flu}
\alias{flu_dictionary}
\title{Protein Sequences of Influenza B Virus Strains}
\format{
A tibble with 130560 rows and 16 variables:
\describe{

\item{\code{protein}}{[chr] Abbreviation of viral protein}
\item{\code{sequence_accession}}{[chr] Unique identifier given to the protein sequence record to allow for tracking of different versions of that sequence.}
\item{\code{complete_genome}}{[chr] Is the viral strain's complete genome known?}
\item{\code{complete_sequence}}{[chr] Is the complete sequence of this viral protein known?}
\item{\code{segment}}{[dbl] One of eight single-stranded RNA segments that encodes the viral protein}
\item{\code{segment_length}}{[dbl] Number of RNA nucleotides in segment}
\item{\code{collection_date}}{[date] Date of sample collection}
\item{\code{host_species}}{[chr] Species that the viral strain infects}
\item{\code{country}}{[chr] Country of strain origin}
\item{\code{state_province}}{[chr] State or province of origin if applicable}
\item{\code{geographic_grouping}}{[chr] Geographic origin of viral strain}
\item{\code{flu_season}}{[chr] For geographic regions in the northern hemisphere, the two digit year for the fall and winter season when the strain was recorded.}
\item{\code{strain_name}}{[chr] Name of viral strain}
\item{\code{sequence}}{[chr] Protein sequence in amino acid }
\item{\code{submission_date}}{[date] Date of entry submission}
\item{\code{passage_history}}{[chr] An indicator of what cell line was used for culturing the virus. Nomenclature for passage history is notoriously unstandardized. See https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6599686/}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://FAIRsharing.org}: IRD; Influenza Research Database; DOI: \url{https://doi.org/10.25504/FAIRsharing.ws7cgw}; Last edited: March 30, 2020, 1:15 p.m.; Last accessed: May 27 2021 4:25 p.m.
}
\usage{
flu

flu_dictionary
}
\description{
A data set of proteins, their sequences, and other attributes of 15,091 unique strains of Influenza B.
}
\section{Functions}{
\itemize{
\item \code{flu_dictionary}: The \code{flu} data dictionary

}}
\references{
Influenza Research Database: An integrated bioinformatics resource for influenza virus research. Zhang Y,Aevermann BD,Anderson TK,Burke DF,Dauphin G,Gu Z,He S,Kumar S,Larsen CN,Lee AJ,Li X,Macken C,Mahaffey C,Pickett BE,Reardon B,Smith T,Stewart L,Suloway C,Sun G,Tong L,Vincent AL,Walters B,Zaremba S,Zhao H,Zhou L,Zmasek C,Klem EB,Scheuermann RH; Nucleic Acids Res ; 2016; 10.1093/nar/gkw857 ;
}
\concept{flu datasets}
\keyword{datasets}
