% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_md.R
\docType{data}
\name{fred_md}
\alias{fred_md}
\alias{fred_md_dictionary}
\title{107 macro-economic indicators, from the FRED-MD database}
\format{
A tibble with 77896 rows and 3 variables:
\describe{

\item{\code{timestamp}}{[double] Date}
\item{\code{series_name}}{[character] Name of the macro-economic indicator}
\item{\code{series_value}}{[double] Value of the indicator}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4654833#.YSQGh9NKg8M}
}
\usage{
fred_md

fred_md_dictionary
}
\description{
A dataset containing various macro-economic indicators tracked by the Federal
Reserve Bank, extracted from the FRED-MD database.
}
\section{Functions}{
\itemize{
\item \code{fred_md_dictionary}: The \code{fred_md} data dictionary
}}

\keyword{datasets}
