% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gshs.R
\docType{data}
\name{gshs}
\alias{gshs}
\alias{gshs_dictionary}
\title{Global School-Based Student Health Survey, Buenos Aires, 2018}
\format{
A tibble with 2254 rows and 11 variables:
\describe{

\item{\code{record}}{[double] Record number}
\item{\code{age}}{[integer] How old are you?}
\item{\code{sex}}{[integer] What is your sex?}
\item{\code{weight}}{[integer] Whether the respondent was determined to be underweight, average weight, or overweight}
\item{\code{hungry}}{[integer] During the past 30 days, how often did you go hungry because there was not enough food in your home?}
\item{\code{fast_food}}{[integer] During the past 7 days, on how many days did you eat food from a fast food restaurant?}
\item{\code{fruit}}{[integer] During the past 7 days, how many times did you eat fruit?}
\item{\code{vegetable}}{[integer] During the past 7 days, how many times did you eat vegetables?}
\item{\code{soft_drink}}{[integer] During the past 7 days, how many times did you drink a can, bottle, or glass of a carbonated soft drink?}
\item{\code{cigarette}}{[integer] During the past 30 days, on how many days did you smoke cigarettes?}
\item{\code{alcohol}}{[integer] During the past 30 days, on how many days did you have at least one drink containing alcohol?}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://extranet.who.int/ncdsmicrodata/index.php/catalog/866/study-description}
}
\usage{
gshs

gshs_dictionary
}
\description{
From \href{https://extranet.who.int/ncdsmicrodata/index.php/catalog/866/study-description}{WHO}:

The GSHS is a school-based survey which uses a self-administered
questionnaire to obtain data on young people's health behaviour and
protective factors related to the leading causes of morbidity and mortality
among children and adults worldwide.

This is the third GSHS conducted by Argentina.
\subsection{Sampling procedure}{

A two-stage cluster sample design was used to produce data representative of
all students in 8 primary/polimodal 1st high - 12 polimodal 5th high in
Argentina. At the first stage, schools were selected with probability
proportional to enrollment size. At the second stage, classes were randomly
selected and all students in selected classes were eligible to participate.
}

\subsection{Response rate}{

The school response rate was 86\%, the student response rate was 74\%, and the
overall response rate was 63\%. A total of 56,981 students participated in the
Argentina GSHS.

This dataset includes only respondents from Buenos Aires, the capital and
largest city in Argentina.
}
}
\section{Functions}{
\itemize{
\item \code{gshs_dictionary}: The \code{gshs} data dictionary
}}

\concept{GSHS datasets}
\keyword{datasets}
