% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital.R
\docType{data}
\name{hospital}
\alias{hospital}
\alias{hospital_dictionary}
\title{Monthly counts of patients use of medical products}
\format{
A tibble with 64428 rows and 5 variables:
\describe{

\item{\code{sku}}{[chr] Hospital stock-keeping unit (SKU) code, representing a specific medical product}
\item{\code{entity_code}}{[chr] Code related to medical product for use with medical billing and insurance purposes}
\item{\code{month}}{[int] Month of interest}
\item{\code{year}}{[int] Year of interest}
\item{\code{patient_counts}}{[int] Number of patients who received the medical product}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://robjhyndman.com/expsmooth/}
}
\usage{
hospital

hospital_dictionary
}
\description{
Monthly patient count for products that are related to medical problems. There are 767 time series that had a mean count of at least 10 and no zeros.
}
\section{Functions}{
\itemize{
\item \code{hospital_dictionary}: The \code{hospital} data dictionary

}}
\examples{
# Convert to `tsibble`
library(tsibble)
library(dplyr)
hospital \%>\%
  mutate(
    date = yearmonth(paste(year, month, sep = "-")),
    .keep = "unused"
  ) \%>\%
  as_tsibble(key = c(sku, entity_code), index = date)
}
\concept{hospital timeseries datasets}
\keyword{datasets}
