% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\docType{data}
\name{lb}
\alias{lb}
\alias{lb_dictionary}
\title{SDTM Lab Test Results data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{VISIT}}{[character] Protocol defined text description of the visit}
\item{\code{VISITNUM}}{[double] Visit number}
\item{\code{LBDT}}{[double] Date of specimen collection}
\item{\code{LBDY}}{[double] Study day of specimen collection}
\item{\code{LBCAT}}{[character] Category of the lab test performed}
\item{\code{LBTEST}}{[character] Name of the lab test performed}
\item{\code{LBSTRESU}}{[character] Standard units for lab test result (LBSTRESN)}
\item{\code{LBSTRESN}}{[double] Lab test result in standard units}
\item{\code{LBSTNRLO}}{[double] Lower limit of normal range lab test result}
\item{\code{LBSTNRHI}}{[double] Upper (high) limit of normal range for lab test result}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
lb

lb_dictionary
}
\description{
\code{lb} describes the results of lab tests performed on subjects in
a simulated clinical trial.
}
\section{Functions}{
\itemize{
\item \code{lb_dictionary}: The \code{lb} data dictionary
}}

\concept{Rho SDTM datasets}
\keyword{datasets}
