% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrd.R
\docType{data}
\name{mdrd_supplemental}
\alias{mdrd_supplemental}
\title{Modification of Diet In Renal Disease - Supplemental Data}
\format{
A tibble with 255 rows and 5 variables:
\describe{

\item{\code{ptid}}{[double] Patient identifier}
\item{\code{sex}}{[character] Sex}
\item{\code{age}}{[double] Age (years)}
\item{\code{height}}{[double] Height (meters)}
\item{\code{weight}}{[double] Weight (kilograms)}
}
}
\source{
\url{https://repository.niddk.nih.gov/studies/mdrd/}
}
\usage{
mdrd_supplemental
}
\description{
Supplemental data for the \link{mdrd} dataset. Note: this data is simulated and is
not from the original MDRD study.
}
\seealso{
Other MDRD datasets: 
\code{\link{mdrd}}
}
\concept{MDRD datasets}
\keyword{datasets}
