% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_visits.R
\docType{data}
\name{medications}
\alias{medications}
\alias{medications_dictionary}
\title{Synthea Synthetic Medications data}
\format{
A spec_tbl_df:
\describe{

\item{\code{start}}{[dttm] The date and time the medication was prescribed.}
\item{\code{stop}}{[dttm] The date and time the prescription ended, if applicable.}
\item{\code{patient}}{[chr] Foreign key to the Patient.}
\item{\code{payer}}{[chr] Foreign key to the Payer.}
\item{\code{encounter}}{[chr] Foreign key to the Encounter where the medication was prescribed.}
\item{\code{code}}{[dbl] Medication code from RxNorm.}
\item{\code{description}}{[chr] Description of the medication.}
\item{\code{base_cost}}{[dbl] The line item cost of the medication.}
\item{\code{payer_coverage}}{[dbl] The amount covered or reimbursed by the Payer.}
\item{\code{dispenses}}{[dbl] The number of times the prescription was filled.}
\item{\code{totalcost}}{[dbl] The total cost of the prescription, including all dispenses.}
\item{\code{reasoncode}}{[dbl] Diagnosis code from SNOMED-CT specifying why this medication was prescribed.}
\item{\code{reasondescription}}{[chr] Description of the reason code.}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://synthea.mitre.org/downloads}
}
\usage{
medications

medications_dictionary
}
\description{
\code{medications} describes simulated medication history of the
patient population
}
\section{Functions}{
\itemize{
\item \code{medications_dictionary}: The \code{medications} data dictionary
}}

\seealso{
Other Synthea Synthetic Patient Population data: 
\code{\link{encounters}}
}
\concept{Synthea Synthetic Patient Population data}
\keyword{datasets}
