% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes_dermatology}
\alias{nhanes_dermatology}
\alias{nhanes_dermatology_dictionary}
\title{National Health and Nutrition Examination Survey, Dermatology, 2017-2018}
\format{
A tibble with 3419 rows and 8 variables:
\describe{

\item{\code{seq_no}}{[double] Respondent sequence number}
\item{\code{sun_reaction}}{[integer] If after several months of not being in the sun, you then went out in the sun without sunscreen or protective clothing for a half hour, which one of these would happen to your skin?}
\item{\code{shade}}{[integer] When you go outside on a very sunny day, for more than one hour, how often do you stay in the shade?}
\item{\code{long_sleeves}}{[integer] When you go outside on a very sunny day, for more than one hour, how often do you wear a long sleeved shirt?}
\item{\code{sunscreen}}{[integer] When you go outside on a very sunny day, for more than one hour, how often do you use sunscreen?}
\item{\code{sunburns}}{[double] How many times in the past year have you had a sunburn?}
\item{\code{time_outdoors_workday}}{[double] During the past 30 days, how much time did you usually spend outdoors between 9 in the morning and 5 in the afternoon on the days that you worked or went to school?}
\item{\code{time_outdoors_weekend}}{[double] During the past 30 days, how much time did you usually spend outdoors between 9 in the morning and 5 in the afternoon on the days when you were not working or going to school?}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Questionnaire&Cycle=2017-2018}
}
\usage{
nhanes_dermatology

nhanes_dermatology_dictionary
}
\description{
From the \href{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/overview.aspx?BeginYear=2017}{NCHS}:

The National Center for Health Statistics (NCHS), Division of Health and
Nutrition Examination Surveys (DHANES), part of the Centers for Disease
Control and Prevention (CDC), has conducted a series of health and nutrition
surveys since the early 1960's. The National Health and Nutrition Examination
Surveys (NHANES) were conducted on a periodic basis from 1971 to 1994. In
1999, NHANES became continuous. Every year, approximately 5,000 individuals
of all ages are interviewed in their homes and complete the health
examination component of the survey. The health examination is conducted in a
mobile examination center (MEC); the MEC provides an ideal setting for the
collection of high quality data in a standardized environment.

The dermatology questionnaire section provides personal interview data on sun
exposure and sun protective behavior.
}
\section{Functions}{
\itemize{
\item \code{nhanes_dermatology_dictionary}: The \code{nhanes_dermatology} data dictionary
}}

\seealso{
Other NHANES datasets: 
\code{\link{nhanes_sleep}}
}
\concept{NHANES datasets}
\keyword{datasets}
