% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes_sleep}
\alias{nhanes_sleep}
\alias{nhanes_sleep_dictionary}
\title{National Health and Nutrition Examination Survey, Sleep Disorders, 2017-2018}
\format{
A tibble with 3419 rows and 8 variables:
\describe{

\item{\code{seq_no}}{[double] Respondent sequence number}
\item{\code{sun_reaction}}{[integer] }
\item{\code{shade}}{[integer] }
\item{\code{long_sleeves}}{[integer] }
\item{\code{sunscreen}}{[integer] }
\item{\code{sunburns}}{[double] }
\item{\code{time_outdoors_workday}}{[double] }
\item{\code{time_outdoors_weekend}}{[double] }
}

An object of class \code{list} of length 3.
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Questionnaire&Cycle=2017-2018}
}
\usage{
nhanes_sleep

nhanes_sleep_dictionary
}
\description{
From the \href{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/overview.aspx?BeginYear=2017}{NCHS}:

The National Center for Health Statistics (NCHS), Division of Health and
Nutrition Examination Surveys (DHANES), part of the Centers for Disease
Control and Prevention (CDC), has conducted a series of health and nutrition
surveys since the early 1960's. The National Health and Nutrition Examination
Surveys (NHANES) were conducted on a periodic basis from 1971 to 1994. In
1999, NHANES became continuous. Every year, approximately 5,000 individuals
of all ages are interviewed in their homes and complete the health
examination component of the survey. The health examination is conducted in a
mobile examination center (MEC); the MEC provides an ideal setting for the
collection of high quality data in a standardized environment.

The sleep disorders (variable name prefix SLQ) data set has questions on
sleep habits and disorders.
}
\section{Functions}{
\itemize{
\item \code{nhanes_sleep_dictionary}: The \code{nhanes_sleep} data dictionary
}}

\seealso{
Other NHANES datasets: 
\code{\link{nhanes_dermatology}}
}
\concept{NHANES datasets}
\keyword{datasets}
