% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ny_air.R
\docType{data}
\name{ny_air}
\alias{ny_air}
\alias{ny_air_dictionary}
\title{Daily predicted New York air quality}
\format{
A tibble with 22692 rows and 5 variables:
\describe{

\item{\code{date}}{[double] Date}
\item{\code{county}}{[character] County FIPS}
\item{\code{county_name}}{[character] County name}
\item{\code{pm25_max}}{[double] Maximum predicted value of PM2.5 concentration}
\item{\code{pm25_median}}{[double] Median predicted value of PM2.5 concentration}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://data.cdc.gov/Environmental-Health-Toxicology/Daily-PM2-5-Concentrations-All-County-2001-2016/7vdq-ztk9}
}
\usage{
ny_air

ny_air_dictionary
}
\description{
Daily predictions of PM2.5 (inhalable particulate matter) concentrations for
New York state in 2016, by county.
}
\section{Functions}{
\itemize{
\item \code{ny_air_dictionary}: The \code{ny_air} data dictionary
}}

\keyword{datasets}
