% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oikolab_weather.R
\docType{data}
\name{oikolab_weather}
\alias{oikolab_weather}
\alias{oikolab_weather_dictionary}
\title{Monash University Weather Data}
\format{
A tibble with 800456 rows and 4 variables:
\describe{

\item{\code{timestamp}}{[dttm] Datetime of observation}
\item{\code{series_name}}{[chr] Name of the climate observation}
\item{\code{type}}{[chr] Type of observation. One of: temperature (C), dewpoint temperature (C), wind speed (m/s), mean sea level pressure (Pa), relative humidity (0-1), surface solar radiation (W/m^2), surface thermal radiation (W/m^2) or total cloud cover (0-1)}
\item{\code{series_value}}{[dbl] Value of the climate observation}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/5184708#.YSVNgNNKiM9}
}
\usage{
oikolab_weather

oikolab_weather_dictionary
}
\description{
Eight time series representing the hourly climate data nearby
Monash University, Clayton, Victoria, Australia from 2010-01-01 to 2021-05-31.
}
\section{Functions}{
\itemize{
\item \code{oikolab_weather_dictionary}: The \code{oikolab_weather} data dictionary
}}

\keyword{datasets}
