% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrian_counts.R
\docType{data}
\name{pedestrian_counts}
\alias{pedestrian_counts}
\alias{pedestrian_counts_dictionary}
\title{Melbourne Pedestrian Counts}
\format{
A tibble with 3132346 rows and 3 variables:
\describe{

\item{\code{timestamp}}{[dttm] Date-time of sensor reading}
\item{\code{series_name}}{[chr] Sensor ID}
\item{\code{series_value}}{[dbl] Hourly count of pedestrians}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656626#.YSVj8FNKjUI}
}
\usage{
pedestrian_counts

pedestrian_counts_dictionary
}
\description{
Hourly pedestrian counts captured from 66 sensors in Melbourne city
starting from May 2009.
}
\section{Functions}{
\itemize{
\item \code{pedestrian_counts_dictionary}: The \code{pedestrian_counts} data dictionary
}}

\keyword{datasets}
