% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rideshare.R
\docType{data}
\name{rideshare}
\alias{rideshare}
\alias{rideshare_dictionary}
\title{Hourly Summaries of Rideshare Service Data}
\format{
A tibble with 84396 rows and 19 variables:
\describe{

\item{\code{source_location}}{[chr] Starting point of the ride}
\item{\code{provider_name}}{[chr] Rideshare service provider}
\item{\code{provider_service}}{[chr] Provider-specific ride type identifier}
\item{\code{timestamp}}{[dttm] Hour}
\item{\code{price_min}}{[dbl] Minimum price estimate for rides in USD}
\item{\code{price_mean}}{[dbl] Mean price estimate for rides in USD}
\item{\code{price_max}}{[dbl] Maxmimum price estimate for rides in USD}
\item{\code{distance_min}}{[dbl] Minimum Distance between source and destination}
\item{\code{distance_mean}}{[dbl] Mean Distance between source and destination}
\item{\code{distance_max}}{[dbl] Maximum Distance between source and destination}
\item{\code{surge_min}}{[dbl] Minimum multiplier by which price was increased, default 1}
\item{\code{surge_mean}}{[dbl] Mean multiplier by which price was increased, default 1}
\item{\code{surge_max}}{[dbl] Maximum multiplier by which price was increased, default 1}
\item{\code{api_calls}}{[dbl] Number of API calls in the hour}
\item{\code{temp}}{[dbl] Temperature (F)}
\item{\code{rain}}{[dbl] Rain in the last hour (inches)}
\item{\code{humidity}}{[dbl] Humidity (\%)}
\item{\code{clouds}}{[dbl] Cloud cover (0-1)}
\item{\code{wind}}{[dbl] Wind speed (mph)}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/5122114#.YS0--9NKiM9}
}
\usage{
rideshare

rideshare_dictionary
}
\description{
This dataset contains various hourly time series representations of
attributes related to Uber and Lyft rideshare services for various locations
in New York between 2018-11-26 and 2018-12-18.
For a given starting location, provider and service,
}
\section{Functions}{
\itemize{
\item \code{rideshare_dictionary}: The \code{rideshare} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
rideshare \%>\%
  as_tsibble(key = c("source_location", "provider_name", "provider_service"), index = timestamp)
}
\keyword{datasets}
