% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverflow.R
\docType{data}
\name{riverflow}
\alias{riverflow}
\alias{riverflow_dictionary}
\title{Daily flow rate of the Saugeen River}
\format{
A tibble with 23741 rows and 2 variables:
\describe{

\item{\code{date}}{[date] Date of observation}
\item{\code{flow_rate}}{[dbl] Volumetric flow rate, in cubic meters per second}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656058#.YSY4VtNKjJJ}
}
\usage{
riverflow

riverflow_dictionary
}
\description{
This dataset contains a time series representing the daily mean flow in cubic
meters per second of the Saugeen River in Ontario, Canada from 01/01/1915
to 12/31/1979.
}
\section{Functions}{
\itemize{
\item \code{riverflow_dictionary}: The \code{riverflow} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
riverflow \%>\%
  tsibble::as_tsibble()
}
\keyword{datasets}
