% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_adverse_events}
\alias{sdtm_adverse_events}
\alias{sdtm_adverse_events_dictionary}
\title{SDTM Adverse Events data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{AESEQ}}{[double] Sequence Number}
\item{\code{AESTDT}}{[double] Start date of the adverse event}
\item{\code{AESTDY}}{[double] The number of days the subject had been enrolled in the study when the adverse event began.}
\item{\code{AEENDT}}{[double] End date of the adverse event}
\item{\code{AEENDY}}{[double] The number of days the subject had been enrolled in the study when the adverse event ended.}
\item{\code{AETERM}}{[character] The reported term for the adverse event, i.e. how the reporter described the adverse event.}
\item{\code{AEDECOD}}{[character] The official term for the adverse event, i.e. the dictionary derived description of the adverse event.}
\item{\code{AEBODSYS}}{[character] Body system or organ class involved in the adverse event.}
\item{\code{AESER}}{[character] Was the adverse event serious? \code{Y} = Yes, \code{N} = No.}
\item{\code{AEONGO}}{[character] Is the adverse event ongoing? \code{Y} = Yes, \code{N} = No.}
\item{\code{AESEV}}{[character] Severity of the adverse event.}
\item{\code{AEREL}}{[character] The investigator's opinion as to whether or not the adverse event was related to the study treatment.}
\item{\code{AEOUT}}{[character] Outcome of the adverse event.}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sdtm_adverse_events

sdtm_adverse_events_dictionary
}
\description{
\code{sdtm_adverse_events} describes adverse event data related to a simulated
clinical trial. An adverse event is an undesirable medical occurrence that
happens while a subject is enrolled in a clinical trial. These occurrences
must be reported and analyzed to see if they are caused by the treatment
under study.
}
\section{Functions}{
\itemize{
\item \code{sdtm_adverse_events_dictionary}: The \code{sdtm_adverse_events} data dictionary
}}

\seealso{
Other Rho SDTM datasets: 
\code{\link{sdtm_concomitant_meds}},
\code{\link{sdtm_demographics}},
\code{\link{sdtm_lab_results}},
\code{\link{sdtm_subject_visits}},
\code{\link{sdtm_vital_signs}}
}
\concept{Rho SDTM datasets}
\keyword{datasets}
