% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_concomitant_meds}
\alias{sdtm_concomitant_meds}
\alias{sdtm_concomitant_meds_dictionary}
\title{SDTM Concomitant Medication data}
\format{
A tibble with 307 rows and 12 variables:
\describe{

\item{\code{USUBJID}}{[chr] Unique Subject Identifier}
\item{\code{CMSEQ}}{[dbl] Sequence Number}
\item{\code{CMSTDT}}{[date] Start date of the concomitant medication}
\item{\code{CMSTDY}}{[dbl] The number of days the subject had been enrolled in the study when the concomitant medication began.}
\item{\code{CMENDT}}{[date] End date of the concomitant medication}
\item{\code{CMENDY}}{[dbl] The number of days the subject had been enrolled in the study when the concomitant medication ended.}
\item{\code{CMTRT}}{[chr] Reported name of the concomitant medication, i.e concomitant treatment}
\item{\code{PREFTERM}}{[chr] The nonproprietary, i.e. generic, name of the concomitant medication}
\item{\code{ATCTEXT2}}{[chr] Anatomical Therapeutic Chemical (ATC) of the concomitant medication}
\item{\code{CMONGO}}{[chr] Is the concomitant medication ongoing? \code{Y} = Yes, \code{N} = No}
\item{\code{CMDOSE}}{[dbl] Amount of concomitant medication taken per administration (dose)}
\item{\code{CMROUTE}}{[chr] Route of administration of the concomitant medication}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sdtm_concomitant_meds

sdtm_concomitant_meds_dictionary
}
\description{
\code{sdtm_concomitant_meds} describes concomitant medication data related to a
simulated clinical trial. Concomitant medications are medications that
subjects take during the trial, concomitant with the treatment being
studied. Concomitant medications are not part of the treatment being
studied but may be confounding variables, or cause interaction effects or
adverse events.
}
\section{Functions}{
\itemize{
\item \code{sdtm_concomitant_meds_dictionary}: The \code{sdtm_concomitant_meds} data dictionary

}}
\seealso{
Other Rho SDTM datasets: 
\code{\link{sdtm_adverse_events}},
\code{\link{sdtm_demographics}},
\code{\link{sdtm_lab_results}},
\code{\link{sdtm_subject_visits}},
\code{\link{sdtm_vital_signs}}
}
\concept{Rho SDTM datasets}
\keyword{datasets}
