% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\docType{data}
\name{solar}
\alias{solar}
\alias{solar_dictionary}
\title{Solar Power}
\format{
A tibble with 7124 rows and 3 variables:
\describe{

\item{\code{plant}}{[chr] ID of the photovoltaic (PV) power plant}
\item{\code{date}}{[date] Date}
\item{\code{power}}{[dbl] Solar power capacity, in megawatts (MW)}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656151#.YSktINNKjSU}
}
\usage{
solar

solar_dictionary
}
\description{
This dataset contains weekly solar power production records from 137
photovoltaic (PV) power plants in Alabama, in 2006. The data were originally
collected by the National Renewable Energy Laboratory (NREL): https://www.nrel.gov/.
}
\section{Functions}{
\itemize{
\item \code{solar_dictionary}: The \code{solar} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
solar \%>\%
  tsibble::as_tsibble(key = plant, index = date)
}
\keyword{datasets}
