% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunspots.R
\docType{data}
\name{sunspots}
\alias{sunspots}
\alias{sunspots_dictionary}
\title{Daily historical sunspot data}
\format{
A tibble with 73924 rows and 2 variables:
\describe{

\item{\code{date}}{[date] Date of observation}
\item{\code{sunspots}}{[int] Observed number of sunspots}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4654773#.YSVlvdNKg8M}
}
\usage{
sunspots

sunspots_dictionary
}
\description{
A dataset containing daily sunspot counts, from
January 08, 1818 to
May 31, 2020.
}
\section{Functions}{
\itemize{
\item \code{sunspots_dictionary}: The \code{sunspots} data dictionary

}}
\examples{
# Convert to `tsibble`
tsibble::as_tsibble(sunspots, key = NULL, index = date)

}
\keyword{datasets}
