% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\docType{data}
\name{sv}
\alias{sv}
\alias{sv_dictionary}
\title{SDTM Subject Visits data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{VISIT}}{[character] Protocol defined text description of the visit}
\item{\code{VISITNUM}}{[double] Visit number}
\item{\code{SVDT}}{[double] Subject visit date}
\item{\code{SVDY}}{[double] Study day of subject visit}
\item{\code{SVSTATUS}}{[character] Status of subject visit: Completed, Terminated, Missed, Expected, Overdue, Failed}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sv

sv_dictionary
}
\description{
\code{sv} describes the clinical visits of subjects in
a simulated clinical trial.
}
\section{Functions}{
\itemize{
\item \code{sv_dictionary}: The \code{sv} data dictionary
}}

\concept{Rho SDTM datasets}
\keyword{datasets}
