% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourism.R
\docType{data}
\name{tourism}
\alias{tourism}
\alias{tourism_dictionary}
\title{Kaggle Tourism forecasting competition time series}
\format{
A tbl_ts:
\describe{

\item{\code{timestamp}}{[date] Month}
\item{\code{series}}{[fct] Series number}
\item{\code{value}}{[dbl] Value of the series during the specified month}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656096}
}
\usage{
tourism

tourism_dictionary
}
\description{
366 monthly time series used in the Kaggle Tourism forecasting competition
}
\section{Functions}{
\itemize{
\item \code{tourism_dictionary}: The \code{tourism} data dictionary
}}

\keyword{datasets}
