% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_births.R
\docType{data}
\name{us_births}
\alias{us_births}
\alias{us_births_dictionary}
\title{us_births}
\format{
A tibble with 7305 rows and 2 variables:
\describe{

\item{\code{timestamp}}{[date] Day}
\item{\code{births}}{[int] Number of births in the US}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://forecastingdata.org/}
}
\usage{
us_births

us_births_dictionary
}
\description{
This dataset contains a single very long daily time series representing the
number of births in US from 01/01/1969 to 31/12/1988. It was extracted from R
mosaicData package. The length of this time series is 7305.

For more details, please refer to Pruim, R., Kaplan, D., Horton, N., 2020.
mosaicData: project mosaic data sets. R package version 0.18.0.
\url{https://CRAN.R-project.org/package=mosaicData}
}
\section{Functions}{
\itemize{
\item \code{us_births_dictionary}: The \code{us_births} data dictionary

}}
\examples{
# Convert to `tsibble`
library(tsibble)
us_births \%>\%
  as_tsibble()
}
\keyword{datasets}
