% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccines.R
\docType{data}
\name{vaccines}
\alias{vaccines}
\alias{vaccines_dictionary}
\title{COVID-19 Vaccine Distribution Allocations}
\format{
A tibble with 3591 rows and 5 variables:
\describe{

\item{\code{jurisdiction}}{[chr] State or territory}
\item{\code{vaccine}}{[chr] Vaccine type}
\item{\code{week}}{[date] Week the vaccines were allocated}
\item{\code{first_dose}}{[dbl] Number of first doses allocated}
\item{\code{second_dose}}{[dbl] Number of second doses allocated}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://data.cdc.gov/browse?category=Vaccinations}
}
\usage{
vaccines

vaccines_dictionary
}
\description{
A dataset containing the number of COVID-19 vaccines allocated in the US by date,
manufacturer, and state/territory.
}
\section{Functions}{
\itemize{
\item \code{vaccines_dictionary}: The \code{vaccines} data dictionary

}}
\keyword{datasets}
