% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.R
\docType{data}
\name{vs}
\alias{vs}
\alias{vs_dictionary}
\title{SDTM Vital Signs data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{VISIT}}{[character] Protocol defined text description of the visit}
\item{\code{VISITNUM}}{[double] Visit number}
\item{\code{VSDT}}{[double] Date vital signs were collected}
\item{\code{VSDY}}{[double] Day of study when vital signs were collected}
\item{\code{VSCAT}}{[character] Category of vital signs test}
\item{\code{VSTEST}}{[character] Name of vital signs test}
\item{\code{VSSTRESU}}{[character] Standard units of vital signs test result}
\item{\code{VSSTRESN}}{[double] Result of vital signs test in standard units}
\item{\code{VSSTNRLO}}{[double] Lower limit of normal range for result of vital signs test}
\item{\code{VSSTNRHI}}{[double] Upper (high) limit of normal range for result of vital signs test}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
vs

vs_dictionary
}
\description{
\code{vs} describes the vital signs of subjects in a simulated
clinical trial. Vital signs include things like heart rate, blood pressure,
temperature, and so on.
}
\section{Functions}{
\itemize{
\item \code{vs_dictionary}: The \code{vs} data dictionary
}}

\concept{Rho SDTM datasets}
\keyword{datasets}
