% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_md.R
\docType{data}
\name{fred_md}
\alias{fred_md}
\alias{fred_md_dictionary}
\title{U.S. macro-economic indicators from the FRED-MD database.}
\format{
A tibble with 751 rows and 7 variables:
\describe{

\item{\code{date}}{[date] Date}
\item{\code{rpi}}{[dbl] Real personal income, in billions of dollars}
\item{\code{hwi}}{[dbl] Help-wanted index: the number of help-wanted ads in major newspapers}
\item{\code{unrate}}{[dbl] Civilian unemployment rate (percent)}
\item{\code{ce16ov}}{[dbl] Thousands of employed civilians}
\item{\code{houst}}{[dbl] Total number of new privately owned houses}
\item{\code{cpiaucsl}}{[dbl] Consumer price index (all items)}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\usage{
fred_md

fred_md_dictionary
}
\description{
A dataset containing six macro-economic indicators tracked by the Federal
Reserve Bank, extracted from the FRED-MD database.

For more information on a variable, look it up by name here:
\url{https://fred.stlouisfed.org/}
}
\section{Functions}{
\itemize{
\item \code{fred_md_dictionary}: The \code{fred_md} data dictionary
}}

\examples{
# Convert to `tsibble`
library(dplyr)
library(tsibble)
fred_md \%>\%
  mutate(date = yearmonth(date)) \%>\%
  as_tsibble(index = date)
}
\keyword{datasets}
