% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes_sleep}
\alias{nhanes_sleep}
\alias{nhanes_sleep_dictionary}
\title{National Health and Nutrition Examination Survey, Sleep Disorders, 2017-2018}
\format{
A tibble with 6161 rows and 11 variables:
\describe{

\item{\code{seq_no}}{[dbl] Respondent sequence number}
\item{\code{sleep_time_workday}}{[time] What time do you usually fall asleep on weekdays or workdays?}
\item{\code{wake_time_workday}}{[time] What time do you usually wake up on weekdays or workdays?}
\item{\code{sleep_length_workday}}{[dbl] Number of hours usually sleep on weekdays or workdays.}
\item{\code{sleep_time_weekend}}{[time] What time do you usually fall asleep on weekends or non-workdays?}
\item{\code{wake_time_weekend}}{[time] What time do you usually wake up on weekends or non-workdays?}
\item{\code{sleep_length_weekend}}{[dbl] Number of hours usually sleep on weekends or non-workdays.}
\item{\code{snore}}{[fct] In the past 12 months, how often did you snore while you were sleeping?}
\item{\code{stop_breathing}}{[fct] In the past 12 months, how often did you snort, gasp, or stop breathing while you were asleep?}
\item{\code{told_doctor}}{[lgl] Have you ever told a doctor or other health professional that you have trouble sleeping?}
\item{\code{overly_sleepy}}{[fct] In the past month, how often did you feel excessively or overly sleepy during the day?}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Questionnaire&Cycle=2017-2018}
}
\usage{
nhanes_sleep

nhanes_sleep_dictionary
}
\description{
From the \href{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/overview.aspx?BeginYear=2017}{NCHS}:

The National Center for Health Statistics (NCHS), Division of Health and
Nutrition Examination Surveys (DHANES), part of the Centers for Disease
Control and Prevention (CDC), has conducted a series of health and nutrition
surveys since the early 1960's. The National Health and Nutrition Examination
Surveys (NHANES) were conducted on a periodic basis from 1971 to 1994. In
1999, NHANES became continuous. Every year, approximately 5,000 individuals
of all ages are interviewed in their homes and complete the health
examination component of the survey. The health examination is conducted in a
mobile examination center (MEC); the MEC provides an ideal setting for the
collection of high quality data in a standardized environment.

The sleep disorders (variable name prefix SLQ) data set has questions on
sleep habits and disorders.
}
\section{Functions}{
\itemize{
\item \code{nhanes_sleep_dictionary}: The \code{nhanes_sleep} data dictionary
}}

\seealso{
Other NHANES datasets: 
\code{\link{nhanes_dermatology}}
}
\concept{NHANES datasets}
\keyword{datasets}
