% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oikolab_weather.R
\docType{data}
\name{oikolab_weather}
\alias{oikolab_weather}
\alias{oikolab_weather_dictionary}
\title{Monash University Weather Data}
\format{
A tibble with 100057 rows and 9 variables:
\describe{

\item{\code{timestamp}}{[dttm] Datetime of observation}
\item{\code{temperature}}{[dbl] temperature (C)}
\item{\code{dewpoint_temperature}}{[dbl] dewpoint temperature (C)}
\item{\code{wind_speed}}{[dbl] wind speed (m/s)}
\item{\code{mean_sea_level_pressure}}{[dbl] mean sea level pressure (Pa)}
\item{\code{relative_humidity}}{[dbl] relative humidity (0-1)}
\item{\code{surface_solar_radiation}}{[dbl] surface solar radiation (W/m^2)}
\item{\code{surface_thermal_radiation}}{[dbl] surface thermal radiation (W/m^2)}
\item{\code{total_cloud_cover}}{[dbl] total cloud cover (0-1)}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/5184708#.YSVNgNNKiM9}
}
\usage{
oikolab_weather

oikolab_weather_dictionary
}
\description{
Eight time series representing the hourly climate data nearby
Monash University, Clayton, Victoria, Australia from 2010-01-01 to 2021-05-31.
}
\section{Functions}{
\itemize{
\item \code{oikolab_weather_dictionary}: The \code{oikolab_weather} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
oikolab_weather \%>\%
  as_tsibble(index = timestamp)
}
\keyword{datasets}
