% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrian_counts.R
\docType{data}
\name{pedestrian_counts}
\alias{pedestrian_counts}
\alias{pedestrian_counts_dictionary}
\title{Melbourne pedestrian count data}
\format{
A tibble with 3132346 rows and 3 variables:
\describe{

\item{\code{date}}{[dttm] Date-time of sensor reading}
\item{\code{sensor_id}}{[chr] Sensor ID}
\item{\code{ped_count}}{[dbl] Hourly count of pedestrians}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656626#.YSVj8FNKjUI}
}
\usage{
pedestrian_counts

pedestrian_counts_dictionary
}
\description{
This dataset records hourly pedestrian counts captured from 66 sensors in
Melbourne city, starting from May 2009.
}
\section{Functions}{
\itemize{
\item \code{pedestrian_counts_dictionary}: The \code{pedestrian_counts} data dictionary
}}

\examples{
# Convert to `tsibble`
library(tsibble)
pedestrian_counts \%>\%
  tsibble::as_tsibble(key = sensor_id, index = date)

}
\keyword{datasets}
