% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_demographics}
\alias{sdtm_demographics}
\alias{sdtm_demographics_dictionary}
\title{SDTM Demographics data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[chr] Unique Subject Identifier}
\item{\code{SITE}}{[chr] Name of study site}
\item{\code{SITEID}}{[chr] Study Site Identifier}
\item{\code{AGE}}{[dbl] Age of subject}
\item{\code{SEX}}{[chr] Sex of subject}
\item{\code{RACE}}{[chr] Race of subject}
\item{\code{ARM}}{[chr] Treatment arm that subject is assigned to}
\item{\code{ARMCD}}{[chr] Code for treatment arm that subject is assigned to}
\item{\code{SBJTSTAT}}{[chr] Subject status in study}
\item{\code{RFSTDTC}}{[date] Reference start date. Usually the day the subject begins treatment.}
\item{\code{RFENDTC}}{[date] Reference end date. Usually the day the subject takes their last treatment.}
\item{\code{RFENDY}}{[dbl] The number of days after the RFSTDTC that the RFENDTC occurred, i.e. the number of days the subject spent in the study/in treatment}
\item{\code{SAFFL}}{[chr] Safety population flag. Did the subject actually receive a treatment (including placebo) and should therefore be included in the population of subjects studied to determine the safety of the treatment? \code{Y} = Yes and \code{N} = No}
\item{\code{SAFFN}}{[dbl] Numeric coding of safety population flag. \code{Y} = Yes and \code{N} = No}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sdtm_demographics

sdtm_demographics_dictionary
}
\description{
\code{sdtm_demographics} describes demographic data related to subjects of a
simulated clinical trial.
}
\section{Functions}{
\itemize{
\item \code{sdtm_demographics_dictionary}: The \code{sdtm_demographics} data dictionary
}}

\seealso{
Other Rho SDTM datasets: 
\code{\link{sdtm_adverse_events}},
\code{\link{sdtm_concomitant_meds}},
\code{\link{sdtm_lab_results}},
\code{\link{sdtm_subject_visits}},
\code{\link{sdtm_vital_signs}}
}
\concept{Rho SDTM datasets}
\keyword{datasets}
