% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_demographics}
\alias{sdtm_demographics}
\alias{sdtm_demographics_dictionary}
\title{SDTM Demographics data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{CMSEQ}}{[double] }
\item{\code{CMSTDT}}{[double] }
\item{\code{CMSTDY}}{[double] }
\item{\code{CMENDT}}{[double] }
\item{\code{CMENDY}}{[double] }
\item{\code{CMTRT}}{[character] }
\item{\code{PREFTERM}}{[character] }
\item{\code{ATCTEXT2}}{[character] }
\item{\code{CMONGO}}{[character] }
\item{\code{CMDOSE}}{[double] }
\item{\code{CMROUTE}}{[character] }
\item{\code{CMINDC}}{[character] }
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sdtm_demographics

sdtm_demographics_dictionary
}
\description{
\code{sdtm_demographics} describes demographic data related to subjects of a
simulated clinical trial.
}
\section{Functions}{
\itemize{
\item \code{sdtm_demographics_dictionary}: The \code{sdtm_demographics} data dictionary
}}

\seealso{
Other Rho SDTM datasets: 
\code{\link{sdtm_adverse_events}},
\code{\link{sdtm_concomitant_meds}},
\code{\link{sdtm_lab_results}},
\code{\link{sdtm_subject_visits}},
\code{\link{sdtm_vital_signs}}
}
\concept{Rho SDTM datasets}
\keyword{datasets}
