% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_lab_results}
\alias{sdtm_lab_results}
\alias{sdtm_lab_results_dictionary}
\title{SDTM Lab Test Results data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[chr] Unique Subject Identifier}
\item{\code{VISIT}}{[chr] Protocol defined text description of the visit}
\item{\code{VISITNUM}}{[dbl] Visit number}
\item{\code{LBDT}}{[date] Date of specimen collection}
\item{\code{LBDY}}{[dbl] Study day of specimen collection}
\item{\code{LBCAT}}{[chr] Category of the lab test performed}
\item{\code{LBTEST}}{[chr] Name of the lab test performed}
\item{\code{LBSTRESU}}{[chr] Standard units for lab test result (LBSTRESN)}
\item{\code{LBSTRESN}}{[dbl] Lab test result in standard units}
\item{\code{LBSTNRLO}}{[dbl] Lower limit of normal range lab test result}
\item{\code{LBSTNRHI}}{[dbl] Upper (high) limit of normal range for lab test result}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sdtm_lab_results

sdtm_lab_results_dictionary
}
\description{
\code{sdtm_lab_results} describes the results of lab tests performed on subjects in
a simulated clinical trial.
}
\section{Functions}{
\itemize{
\item \code{sdtm_lab_results_dictionary}: The \code{sdtm_lab_results} data dictionary
}}

\seealso{
Other Rho SDTM datasets: 
\code{\link{sdtm_adverse_events}},
\code{\link{sdtm_concomitant_meds}},
\code{\link{sdtm_demographics}},
\code{\link{sdtm_subject_visits}},
\code{\link{sdtm_vital_signs}}
}
\concept{Rho SDTM datasets}
\keyword{datasets}
