% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_subject_visits}
\alias{sdtm_subject_visits}
\alias{sdtm_subject_visits_dictionary}
\title{SDTM Subject Visits data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[character] Unique Subject Identifier}
\item{\code{VISIT}}{[character] Protocol defined text description of the visit}
\item{\code{VISITNUM}}{[double] Visit number}
\item{\code{SVDT}}{[double] Subject visit date}
\item{\code{SVDY}}{[double] Study day of subject visit}
\item{\code{SVSTATUS}}{[character] Status of subject visit: Completed, Terminated, Missed, Expected, Overdue, Failed}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sdtm_subject_visits

sdtm_subject_visits_dictionary
}
\description{
\code{sdtm_subject_visits} describes the clinical visits of subjects in
a simulated clinical trial.
}
\section{Functions}{
\itemize{
\item \code{sdtm_subject_visits_dictionary}: The \code{sdtm_subject_visits} data dictionary
}}

\seealso{
Other Rho SDTM datasets: 
\code{\link{sdtm_adverse_events}},
\code{\link{sdtm_concomitant_meds}},
\code{\link{sdtm_demographics}},
\code{\link{sdtm_lab_results}},
\code{\link{sdtm_vital_signs}}
}
\concept{Rho SDTM datasets}
\keyword{datasets}
