% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\docType{data}
\name{sdtm_vital_signs}
\alias{sdtm_vital_signs}
\alias{sdtm_vital_signs_dictionary}
\title{SDTM Vital Signs data}
\format{
A spec_tbl_df:
\describe{

\item{\code{USUBJID}}{[chr] Unique Subject Identifier}
\item{\code{VISIT}}{[chr] Protocol defined text description of the visit}
\item{\code{VISITNUM}}{[dbl] Visit number}
\item{\code{VSDT}}{[date] Date vital signs were collected}
\item{\code{VSDY}}{[dbl] Day of study when vital signs were collected}
\item{\code{VSCAT}}{[chr] Category of vital signs test}
\item{\code{VSTEST}}{[chr] Name of vital signs test}
\item{\code{VSSTRESU}}{[chr] Standard units of vital signs test result}
\item{\code{VSSTRESN}}{[dbl] Result of vital signs test in standard units}
\item{\code{VSSTNRLO}}{[dbl] Lower limit of normal range for result of vital signs test}
\item{\code{VSSTNRHI}}{[dbl] Upper (high) limit of normal range for result of vital signs test}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://rhoinc.github.io/data-library/}
}
\usage{
sdtm_vital_signs

sdtm_vital_signs_dictionary
}
\description{
\code{sdtm_vital_signs} describes the vital signs of subjects in a simulated
clinical trial. Vital signs include things like heart rate, blood pressure,
temperature, and so on.
}
\section{Functions}{
\itemize{
\item \code{sdtm_vital_signs_dictionary}: The \code{sdtm_vital_signs} data dictionary
}}

\seealso{
Other Rho SDTM datasets: 
\code{\link{sdtm_adverse_events}},
\code{\link{sdtm_concomitant_meds}},
\code{\link{sdtm_demographics}},
\code{\link{sdtm_lab_results}},
\code{\link{sdtm_subject_visits}}
}
\concept{Rho SDTM datasets}
\keyword{datasets}
