% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourism.R
\docType{data}
\name{tourism}
\alias{tourism}
\alias{tourism_dictionary}
\title{Kaggle Tourism forecasting competition time series}
\format{
A tibble with 109280 rows and 4 variables:
\describe{

\item{\code{month}}{[date] Start date of the month}
\item{\code{city}}{[chr] Fabricated city}
\item{\code{country}}{[chr] Fabricated country}
\item{\code{tourists}}{[int] Number of tourists who visited the location in the specified month}
}

An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/record/4656096}
}
\usage{
tourism

tourism_dictionary
}
\description{
From \href{https://robjhyndman.com/papers/forecompijf.pdf}{Athanasopoulos, Hyndman, Song, and Wu (2010)}

The data include 366 monthly series. They were supplied by both tourism
bodies (such as Tourism Australia, the Hong Kong Tourism Board and Tourism
New Zealand) and various academics, who had used them in previous tourism
forecasting studies.

In order to adhere to all confidentiality agreements with all parties, the
data are presented under coded titles.

City and countries were fabricated and assigned to the data. This data does
not reflect true tourism trends for these cities.
}
\section{Functions}{
\itemize{
\item \code{tourism_dictionary}: The \code{tourism} data dictionary
}}

\examples{
# Convert to `tsibble`
tourism \%>\%
  dplyr::mutate(month = tsibble::yearmonth(month)) \%>\%
  tsibble::as_tsibble(key = c("city", "country"), index = month)
}
\keyword{datasets}
